Move test directories and update jest config
