+++ server/tests/inputs.test.js
